requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
psycopg2-binary==2.8.6
email-validator==1.1.2
boto3==1.17.10